+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## 0.0.2
+- Correct an issue with path location in the edapack script
+
+- Correct an issue with the index URL in etc/sources
+
+- Enhance the install process to update source indexes first
+
 ## 0.0.1
 
 - Initial trial release of EDAPack. Linux-only, and just a couple of tools.